+++ competition/models.py
@@ -4,11 +4,7 @@ from io import BytesIO
 from operator import itemgetter
 
 import pdf2image
-from base.managers import UnspecifiedValueManager
-from base.models import RestrictedFileField
-from base.utils import mime_type
-from base.validators import phone_number_validator, school_year_validator
-from django.conf import settings
+
 from django.contrib.sites.models import Site
 from django.core.exceptions import ValidationError
 from django.core.files.base import ContentFile
@@ -20,44 +16,14 @@ from django.dispatch import receiver
 from django.shortcuts import get_object_or_404
 from django.utils.functional import cached_property
 from django.utils.timezone import now
-from profile.models import School
 
+
+from base.managers import UnspecifiedValueManager
+from base.models import RestrictedFileField
+from base.utils import mime_type
+from base.validators import school_year_validator
 from competition import utils
-
-
-class Profile(models.Model):
-    class Meta:
-        verbose_name = 'profil'
-        verbose_name_plural = 'profily'
-
-    user = models.OneToOneField(
-        settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
-
-    nickname = models.CharField(
-        verbose_name='prezývka', max_length=32, blank=True, )
-
-    school = models.ForeignKey(
-        School, on_delete=models.SET(School.objects.get_unspecified_value),
-        verbose_name='škola')
-
-    year_of_graduation = models.PositiveSmallIntegerField(
-        verbose_name='rok maturity')
-
-    phone = models.CharField(
-        verbose_name='telefónne číslo', max_length=32, blank=True,
-        validators=[phone_number_validator],
-        help_text='Telefonné číslo v medzinárodnom formáte (napr. +421 123 456 789).')
-
-    parent_phone = models.CharField(
-        verbose_name='telefónne číslo na rodiča', max_length=32, blank=True,
-        validators=[phone_number_validator],
-        help_text='Telefonné číslo v medzinárodnom formáte (napr. +421 123 456 789).')
-
-    gdpr = models.BooleanField(
-        verbose_name='súhlas so spracovaním osobných údajov', default=False)
-
-    def __str__(self):
-        return str(self.user)
+from profile.models import Profile, School
 
 
 class Competition(models.Model):
