+++ profile/models.py
@@ -1,5 +1,7 @@
 from django.db import models
 from base.managers import UnspecifiedValueManager
+from django.conf import settings
+from base.validators import phone_number_validator
 
 
 class County(models.Model):
@@ -68,3 +70,38 @@ class School(models.Model):
     def stitok(self):
         return f'\\stitok{{{ self.name }}}{{{ self.city }}}' \
                f'{{{ self.printable_zip_code }}}{{{ self.street }}}'
+
+
+class Profile(models.Model):
+    class Meta:
+        verbose_name = 'profil'
+        verbose_name_plural = 'profily'
+
+    user = models.OneToOneField(
+        settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
+
+    nickname = models.CharField(
+        verbose_name='prezývka', max_length=32, blank=True, )
+
+    school = models.ForeignKey(
+        School, on_delete=models.SET(School.objects.get_unspecified_value),
+        verbose_name='škola')
+
+    year_of_graduation = models.PositiveSmallIntegerField(
+        verbose_name='rok maturity')
+
+    phone = models.CharField(
+        verbose_name='telefónne číslo', max_length=32, blank=True,
+        validators=[phone_number_validator],
+        help_text='Telefonné číslo v medzinárodnom formáte (napr. +421 123 456 789).')
+
+    parent_phone = models.CharField(
+        verbose_name='telefónne číslo na rodiča', max_length=32, blank=True,
+        validators=[phone_number_validator],
+        help_text='Telefonné číslo v medzinárodnom formáte (napr. +421 123 456 789).')
+
+    gdpr = models.BooleanField(
+        verbose_name='súhlas so spracovaním osobných údajov', default=False)
+
+    def __str__(self):
+        return str(self.user)
