--- school/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class SchoolConfig(AppConfig):
-    name = 'school'
-    verbose_name = 'školy'
