--- school/fixtures/schools_custom.json
@@ -1,18 +1,0 @@
-[
-    {
-        "model": "school.School",
-        "pk": 0,
-        "fields": {
-            "district": 0,
-            "name": "Iná škola"
-        }
-    },
-    {
-        "model": "school.School",
-        "pk": 1,
-        "fields": {
-            "district": 0,
-            "name": "Bez školy"
-        }
-    }
-]
