--- school/fixtures/sources/school_fixtures-requirements.txt
@@ -1,8 +1,0 @@
-numpy==1.18.2
-pandas==1.0.3
-python-dateutil==2.8.1
-pytz==2019.3
-six==1.14.0
-Unidecode==1.1.1
-wget==3.2
-xlrd==1.2.0
