--- school/serializers.py
@@ -1,21 +1,0 @@
-from rest_framework import serializers
-
-from .models import County, District, School
-
-
-class CountySerializer(serializers.ModelSerializer):
-    class Meta:
-        model = County
-        fields = '__all__'
-
-
-class DistrictSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = District
-        fields = '__all__'
-
-
-class SchoolSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = School
-        fields = '__all__'
