--- school/views.py
@@ -1,21 +1,0 @@
-from rest_framework import viewsets
-
-from .models import County, District, School
-from .serializers import SchoolSerializer, CountySerializer, DistrictSerializer
-
-
-class CountyViewSet(viewsets.ReadOnlyModelViewSet):
-    queryset = County.objects.all()
-    serializer_class = CountySerializer
-
-
-class DistrictViewSet(viewsets.ReadOnlyModelViewSet):
-    queryset = District.objects.all()
-    serializer_class = DistrictSerializer
-    filterset_fields = ['county', ]
-
-
-class SchoolViewSet(viewsets.ReadOnlyModelViewSet):
-    queryset = School.objects.all()
-    serializer_class = SchoolSerializer
-    filterset_fields = ['district', ]
